requirements.txt
nltk
geoip2
git+https://github.com/CloudBotIRC/mcstatus.git@master
cleverwrap
future
microdata
sqlalchemy
watchdog
lxml
beautifulsoup4
feedparser
requests
psutil
requests-oauthlib
tweepy
pyenchant
pythonwhois
imgurpython
isodate
python-dateutil
